24 July 2025/mcp-server-db/requirements.txt
aiosqlite
python-dotenv
mcp-server
pydantic
email-validator
mcp-use
langchain-groq